+++ CHANGELOG.md
@@ -9,6 +9,7 @@ and this project adheres to [Semantic Versioning](https://semver.org/spec/v2.0.0
 ### Added
 - Custom logos support in dark mode
 - Missed views rate statistics for sites
+- Filtering by added and accepted dates on banner classification page
 ### Changed
 - Targeting custom option UX - remove unintuitive batch input
 ### Fixed